[HIA-721]
WIP Fixed gateway properly and all the tests pass
