+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/mockservers/RiskManagementApiMockServer.kt
@@ -1,9 +1,9 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.mockservers
 
 import com.github.tomakehurst.wiremock.WireMockServer
-import com.github.tomakehurst.wiremock.client.WireMock.matching
-import com.github.tomakehurst.wiremock.client.WireMock.get
 import com.github.tomakehurst.wiremock.client.WireMock.aResponse
+import com.github.tomakehurst.wiremock.client.WireMock.get
+import com.github.tomakehurst.wiremock.client.WireMock.matching
 import org.springframework.http.HttpStatus
 
 class RiskManagementApiMockServer : WireMockServer(WIREMOCK_PORT) {
@@ -17,7 +17,7 @@ class RiskManagementApiMockServer : WireMockServer(WIREMOCK_PORT) {
     status: HttpStatus = HttpStatus.OK,
   ) {
     stubFor(
-      get("risks/crn/$crn/risk-management-plan")
+      get("/risks/crn/$crn/risk-management-plan")
         .withHeader("Authorization", matching("Bearer ${HmppsAuthMockServer.TOKEN}"))
         .willReturn(
           aResponse()
@@ -27,5 +27,4 @@ class RiskManagementApiMockServer : WireMockServer(WIREMOCK_PORT) {
         ),
     )
   }
-
 }
